Creating the 'Album' class to represent sets of music in the application
Continuing the development of the project, creating the 'Album' class in the 'Album.cs' file to represent sets of songs in our system. Creating the 'Nome' property in the 'Album' class, and the 'DuracaoTotal' property, which will be equal to the sum of the durations of all songs in the album. Creating the private attribute 'musicas,' which will be a list of songs associated with an album. Finally, the 'ExibirMusicas' method was created in the 'Album' class to display the names of the songs present in the album and the 'DuracaoTotal' of the album, and applying these changes in the 'Program.cs' file.