Creating the 'Banda' class, implementing constructors, and refining the relationships between classes
Continuing the project by creating the 'Banda' class to represent bands in the system, with the 'Nome' property, the private field 'albuns' to represent a list of albums associated with this band, the 'AdicionarAlbum' method to add an album to this list of albums, and the 'ExibirDiscografia' method to print information about the band and the albums linked to it. Modifying the classes to implement constructors in these classes, allowing the insertion of information at the time of creating instances of a class. Finally, refining the interactions between the classes, creating the need to link a band when creating a new instance of 'Musica'.